+++ src/client/editions.ts
@@ -6,6 +6,7 @@ import {
 import { ShareIcon } from 'components/editions/shareIcon';
 import { createElement as h } from 'react';
 import ReactDOM from 'react-dom';
+import interactives from './interactives';
 
 initPingEditionsRendering();
 
@@ -56,5 +57,6 @@ const adjustGalleryImages = (): void => {
 
 initLightbox();
 adjustGalleryImages();
+interactives();
 
 ReactDOM.render(h(ShareIcon), document.querySelector('.js-share-button'));
